Re-org ftdi internal interfaces
